Move  handlers and logic to the parent component
