+++ src/App.js
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
 import GameCard from "./components/GameCard/GameCard";
+import { getRandomCards } from "./utils/helpers/App";
 import "./App.scss";
 
 function App() {
@@ -15,39 +16,55 @@ function App() {
     }, 5000);
   };
 
-  const getRandomNumber = (min, max) => {
-    const rand = min - 0.5 + Math.random() * (max - min + 1);
-    return Math.round(rand);
+  const getActiveCards = () => {
+    return cardsData.filter(
+      (card) => card.isHidden === false && card.guessed === false
+    );
   };
 
-  const isUniqueCardValue = (initialCards, newValue) => {
-    return !initialCards.find((card) => card.value === newValue);
+  const onCardClick = (id, isHidden) => {
+    setCardsData(
+      cardsData.map((item) =>
+        item.id === id ? { ...item, isHidden: !isHidden } : item
+      )
+    );
   };
 
-  const getRandomCards = () => {
-    let initialCards = [];
+  const checkIsVisible = (visibleCards, card) => {
+    if (visibleCards.length === 2) {
+      let areIdenticalValues = true;
 
-    while (initialCards.length < 32) {
-      const newValue = getRandomNumber(1, 60);
+      visibleCards.forEach((visibleCard) => {
+        if (visibleCard.value !== card.value) {
+          areIdenticalValues = false;
+        }
+      });
 
-      if (isUniqueCardValue(initialCards, newValue)) {
-        initialCards.push(
-          {
-            value: newValue,
-            isHidden: false,
-            id: initialCards.length,
-            guessed: false,
-          },
-          {
-            value: newValue,
-            isHidden: false,
-            id: initialCards.length + 1,
-            guessed: false,
+      if (areIdenticalValues) {
+        const newCards = cardsData.map((card) => {
+          if (card.isHidden === false) {
+            return {
+              ...card,
+              guessed: true,
+            };
           }
-        );
+          return card;
+        });
+        setCardsData(newCards);
+      }
+
+      if (!areIdenticalValues) {
+        const newCards = cardsData.map((card) => {
+          if (card.isHidden === false && card.guessed === false) {
+            return { ...card, isHidden: true };
+          }
+
+          return card;
+        });
+
+        setTimeout(() => setCardsData(newCards), 500);
       }
     }
-    return initialCards.sort(() => Math.random() - 0.5);
   };
 
   useEffect(() => {
@@ -62,10 +79,11 @@ function App() {
           <GameCard
             key={card.id}
             card={card}
-            isHidden={card.isHidden}
             cardsData={cardsData}
             setCardsData={setCardsData}
-            id={card.id}
+            getActiveCards={getActiveCards}
+            onCardClick={onCardClick}
+            checkIsVisible={checkIsVisible}
           />
         ))}
       </div>
