+++ src/components/GameCard/GameCard.js
@@ -1,61 +1,13 @@
 import { useEffect } from "react";
 import "./GameCard.scss";
 
-const GameCard = ({ isHidden, cardsData, setCardsData, id, card }) => {
-  const onCardClick = () => {
-    setCardsData(
-      cardsData.map((item) =>
-        item.id === id ? { ...item, isHidden: !isHidden } : item
-      )
-    );
-  };
-
-  const getActiveCards = () => {
-    return cardsData.filter(
-      (card) => card.isHidden === false && card.guessed === false
-    );
-  };
-
-  const checkIsVisible = (visibleCards) => {
-    if (visibleCards.length === 2) {
-      let areIdenticalValues = true;
-
-      visibleCards.forEach((visibleCard) => {
-        if (visibleCard.value !== card.value) {
-          areIdenticalValues = false;
-        }
-      });
-
-      if (areIdenticalValues) {
-        const newCards = cardsData.map((card) => {
-          if (card.isHidden === false) {
-            return {
-              ...card,
-              guessed: true,
-            };
-          }
-          return card;
-        });
-        setCardsData(newCards);
-      }
-
-      if (!areIdenticalValues) {
-        const newCards = cardsData.map((card) => {
-          if (card.isHidden === false && card.guessed === false) {
-            return { ...card, isHidden: true };
-          }
-
-          return card;
-        });
-
-        setTimeout(() => setCardsData(newCards), 500);
-      }
-    }
-  };
+const GameCard = ({ checkIsVisible, card, getActiveCards, onCardClick }) => {
+  const {id, isHidden } = card;
 
   const updateCards = () => {
     const activeCards = getActiveCards();
-    checkIsVisible(activeCards);
+
+    checkIsVisible(activeCards, card);
   };
 
   const getCardStyles = () => {
@@ -75,7 +27,7 @@ const GameCard = ({ isHidden, cardsData, setCardsData, id, card }) => {
   return (
     <div
       className={getCardStyles()}
-      onClick={!card.guessed ? onCardClick : () => console.log("guessed")}
+      onClick={!card.guessed ? () => onCardClick(id, isHidden) : () => console.log("guessed")}
     >
       {!isHidden && card.value}
     </div>
